Move Database Connection to Class
The __main__ section makes this file look like it could be imported and subclassed, but the previous implementation would make that difficult. To make it importable, the connection information has been moved to the constructor of the class and has been made an attribute of the class. The database filename has been moved to a parameter to allow subclasses to change the file location. A file property has also been added to allow access to the filename that the database has been created under.

Signed-off-by: XakNitram <xaknitram@gmail.com>